+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `YoutubeButtonIcon` and `YoutubeIcon` components.
+- Fix: Update all social `buttonIcon` sizes.
 - Feature: Add `ButtonImageWithText` component and story.
 
 ## [20.14.0] - 2018-05-29